### 2022-04-27 #4350 @editorialbot
[PRE REVIEW]: NO-DOUBT: A Decision Support Web Platform based on AHP
**Submitting author:** @gdede-hua (<a href="http://orcid.org/0000-0003-2947-7470">Georgia Dede</a>)
**Repository:** https://github.com/gdede-hua/Decision_Survey_Platform
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
