### 2022-04-27 #4350 @danielskatz
👋 @gdede-hua - note that when you want to use multiple citations in one place in your paper, you should use `[@ref1; @ref2]` rather than `[@ref1][@ref2]`. You could fix these in your source, then use the command `@editorialbot generate pdf` to make a new PDF. editorialbot commands need to be the first entry in a new comment.
