### 2022-05-10 #4350 @fabian-s
:wave: @fsteeg & @gcattan & @TimCam

Would any of you be willing to review this submission for JOSS?
