### 2022-05-18 #4350 @fabian-s
:wave: @rreinecke @TimCam @dvysardana

Would any of you be willing to review this submission for JOSS?
