### 2022-05-18 #4403 @fabian-s
```@editorialbot generate my checklist```

**I'm setting the target date for the reviews to June 25 -- 5 weeks from now.**

@gdede-hua, I apologize this will take so long -- JOSS aims for quick turnarounds but it just didn't work out this time.
