### 2022-06-06 gdede-hua/Decision_Survey_Platform#1 @gcattan
[JOSS] Community Guidelines
# Issue https://github.com/openjournals/joss-reviews/issues/4403

Please, add community guidelines to the Github:

> Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
