### 2022-06-06 gdede-hua/Decision_Survey_Platform#2 @gcattan
[JOSS] Add tests and test coverage
# Issue https://github.com/openjournals/joss-reviews/issues/4403

Please, add automated tests and test coverage to this repository:

> Are there automated tests or manual steps described so that the functionality of the software can be verified?
