### 2022-06-06 #4403 @gcattan
Hi,
Why there is two `decision-survey-platform` in your profile? Why there is no commit history on this repository? Why there are so many authors in the paper when you are only two contributors and 6 commits?
