### 2022-06-07 #4403 @gdede-hua
> @gdede-hua Why there is two `decision-survey-platform` in your profile? Why there is no commit history on this repository? Why there are so many authors in the paper when you are only two contributors and 6 commits?

@gcattan Thank you for your feedback. Please find below the answers to your questions:
-There was an older software in PHP named decision-survey-platform.
-There is no commit history because we have squashed the commits in order to have a clear branch for the publication in JOSS.
-The rest authors have contributed to the research of the AHP methodology and the related algorithm but in the the development of the software.
