### 2022-06-07 gdede-hua/Decision_Survey_Platform#2 @gdede-hua
@gcattan Thank you for your feedback.
Please be kindly informed that the manual steps describing the main functionalities are presented in the readme file. Following these steps one may verify the software functionalities. Is that ok?
