+++ README.md
@@ -86,5 +86,7 @@ Below you may find example answers for an assigned research.
 
 ## General Guidelines
 -Third parties that want to contribute to the software may have access to the related files and the source code.
+
 -Anyone who wants to report an issue or a problem with the software may open an issue in the software repository in Github.
+
 -Anyone who seeks for support regarding the installation/usage of the software may open an issue in the software repository in Github.
