### 2022-06-07 gdede-hua/Decision_Survey_Platform#1 @gdede-hua
@gcattan Thank you for your feedback.
Please be kindly informed that the readme file has been updated and further guidelines are included as per your comments.
