### 2022-06-07 #4403 @gdede-hua
> Thanks, @gdede-hua & @gcattan!
>
> * To speed up the review process, I would like to ask @gdede-hua to start discussing/addressing the other issues  identified by @gcattan while we wait for rreinecke's review.
> * Regarding co-authorship, [JOSS's guidelines](https://joss.readthedocs.io/en/latest) say that "[p]urely financial (such as being named on an award) and organizational (such as general supervision of a research group) contributions are not considered sufficient for co-authorship of JOSS submissions, but active project direction and other forms of non-code contributions are. _The authors themselves assume responsibility for deciding who should be credited with co-authorship, and co-authors must always agree to be listed._"  Does the current list of co-authors fulfill these very lenient criteria?

@fabian-s Many thanks for your prompt response.
I confirm that the current list of co-authors fulfill the aforementioned criteria.
