### 2022-06-07 gdede-hua/Decision_Survey_Platform#2 @gcattan
You need to provide evidence that the software works as expected, and to provide a mechanism to guarantee that any further changes will not break its behavior.
if you do not want to write automated tests for this, you can write high-level user stories in Gherkin syntax for example, or provide detailed tutorials that you can use later for functional validation testing.
The pictures in the readme do not provide information on whether or not the computation is _correct_.
