### 2022-06-09 gdede-hua/Decision_Survey_Platform#3 @rreinecke
Example usage: Add an example use-case
You provide a quite extensive intro with screenshots on how to add users etc. but there is no explanation on how the tool  then can then be used to solve a real problem. Please add a (simple) example that showcases the use of the tool.

openjournals/joss-reviews#4403
