### 2022-06-09 gdede-hua/Decision_Survey_Platform#4 @rreinecke
Functionality documentation: add documentation
https://github.com/openjournals/joss-reviews/issues/4403

You nicely describe the use of the software in your paper draft but apart from the readme I was not able to find any proper documentation of the internal software components.

- Please document the API and provide guidance on where which component can be found e.g. you apply a MVC pattern to implement you tool. That should be at least pointed out.
- Provide guidance on how users can modify the software and where contributors may start when they 1) want to add new things to the user interface or 2) modify the internal algorithm
- So far the documentation of the source code is not acceptable. Just picking one file by random: https://github.com/gdede-hua/Decision_Survey_Platform/blob/master/app-maven/src/main/java/com/hua/model/FactorAnswer.java There are no docstrings that tell me what the class FactorAnswer is suppose to do. Please add extensive documentation that allows others to understand and reuse your code: https://www.oracle.com/de/technical-resources/articles/java/javadoc-tool.html
