### 2022-06-09 gdede-hua/Decision_Survey_Platform#5 @rreinecke
Community guidelines
openjournals/joss-reviews/issues/4403

You have a short explanation at the end of your readme. Please implement it as it is best practice and refer to the guidelines in your readme:

https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors
