### 2022-06-15 #4403 @gcattan
Installation with Docker is smooth, but you should indicate a description of your API ports.
Also, if I create an account and log in, I am not sure how to start from there.
I recommend providing a detailed tutorial with ideally a demo docker image, with data already loaded and 2 accounts (one administrator, and one standard user).
