### 2022-06-15 gdede-hua/Decision_Survey_Platform#6 @gcattan
Add demo data and describe API ports
As per: https://github.com/openjournals/joss-reviews/issues/4403#issuecomment-1156833231
