### 2022-09-10 #4403 @arfon
@gdede-hua – are you still interested in pursuing this submission with JOSS?

@fabian-s – it seems like the author may have abandoned this submission, I would suggest giving @gdede-hua one month notice to return here with substantial updates otherwise I think we should reject this submission as abandoned by the author. What do you think?
