### 2022-09-19 #4403 @fabian-s
@gdede-hua are you still interested in pursuing this JOSS publication?
If so please update us on your progress on this and/or give an estimate how much more time you will need to adress the issues pointed out by our reviewers.
