### 2022-09-20 #4403 @gdede-hua
Dear Fabien,

I am interested in pursuing this publication. May I kindly ask you to give
more time (e.g. by the middle of October?)

Many thanks in advance.

On Mon, Sep 19, 2022 at 11:29 AM Fabian Scheipl ***@***.***>
wrote:

> @gdede-hua <https://github.com/gdede-hua> are you still interested in
> pursuing this JOSS publication?
> If so please update us on your progress on this and/or give an estimate
> how much more time you will need to adress the issues pointed out by our
> reviewers.
>
