+++ app-maven/src/main/java/com/hua/ServletInitializer.java
@@ -5,6 +5,10 @@ import org.springframework.boot.web.servlet.support.SpringBootServletInitializer
 
 public class ServletInitializer extends SpringBootServletInitializer {
 
+	/**
+	 * Start up the Spring project
+	 *  @param application - Start up arguments
+	 */
 	@Override
 	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
 		return application.sources(DecisionMakerApplication.class);
