+++ app-maven/src/main/java/com/hua/repository/ProblemsRepository.java
@@ -6,7 +6,9 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.hua.model.Problem;
-
+/**
+ * Interface for the research communication with the database
+ */
 public interface ProblemsRepository extends JpaRepository<Problem, Integer> {
 	public Optional<List<Problem>> findAllByUserUsername(String username);
 }
