+++ app-maven/src/main/java/com/hua/controller/ProblemsController.java
@@ -5,9 +5,6 @@ import java.sql.Date;
 import java.util.ArrayList;
 import java.util.Optional;
 
-import javax.servlet.http.HttpSession;
-
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -34,16 +31,11 @@ import com.hua.util.Mailer;
 @RequestMapping("/problems")
 public class ProblemsController {
 
-	@Autowired
-	private ProblemsRepository problemsRepository;
-	@Autowired
-	private UsersGroupRepository userGroupRepository;
-	@Autowired
-	private ProblemsUserRepository problemsUserRepository;
-	@Autowired
-	private ProblemUserExpireRepository problemUserExpireRepository;
-	@Autowired
-	private UsersRepository userRepository;
+	private final ProblemsRepository problemsRepository;
+	private final UsersGroupRepository userGroupRepository;
+	private final ProblemsUserRepository problemsUserRepository;
+	private final ProblemUserExpireRepository problemUserExpireRepository;
+	private final UsersRepository userRepository;
 	
 	@Value("${security.mail.sslEnable}")
 	private boolean sslEnable;
@@ -57,34 +49,67 @@ public class ProblemsController {
 	private String socketPort;
 	@Value("${security.mail.smtpPort}")
 	private String smtpPort;
-	
-	
+
+	public ProblemsController(ProblemsRepository problemsRepository, UsersGroupRepository userGroupRepository, ProblemsUserRepository problemsUserRepository, ProblemUserExpireRepository problemUserExpireRepository, UsersRepository userRepository) {
+		this.problemsRepository = problemsRepository;
+		this.userGroupRepository = userGroupRepository;
+		this.problemsUserRepository = problemsUserRepository;
+		this.problemUserExpireRepository = problemUserExpireRepository;
+		this.userRepository = userRepository;
+	}
+
+	/**
+	 * Endpoint for the load of the research page
+	 *
+	 * @param model Interface to set up information for the page
+	 * @param principal Information for the login user
+	 *
+	 * @return to research page
+	 */
 	@GetMapping("manage")
 	public String getManage(Model model, Principal principal) {
-		
-		model.addAttribute("problems", problemsRepository.findAllByUserUsername(principal.getName()).isPresent() ? 
-				problemsRepository.findAllByUserUsername(principal.getName()).get() : new ArrayList<>() );
-		model.addAttribute("usersGroup", userGroupRepository.findAllByUserUsername(principal.getName()).isPresent() ?
-				userGroupRepository.findAllByUserUsername(principal.getName()).get() : new ArrayList<>());
-		return "problems/problems.html";
+		return loadProblemsPage(model, principal);
 	}
+	/**
+	 * Endpoint for to delete a research
+	 *
+	 * @param model Interface to set up information for the page
+	 * @param principal Information for the login user
+	 * @param id the id of the research
+	 *
+	 * @return to research page
+	 */
 	@GetMapping("manage/delete/{id}")
 	public String getDelete(Model model, Principal principal, @PathVariable(required = false) String id) {
 		problemsRepository.deleteById(Integer.parseInt(id));
-		model.addAttribute("problems", problemsRepository.findAllByUserUsername(principal.getName()).isPresent() ? 
-				problemsRepository.findAllByUserUsername(principal.getName()).get() : new ArrayList<>() );
-		model.addAttribute("usersGroup", userGroupRepository.findAllByUserUsername(principal.getName()).isPresent() ?
-				userGroupRepository.findAllByUserUsername(principal.getName()).get() : new ArrayList<>());
-		return "problems/problems.html";
[…]
