+++ app-maven/src/main/java/com/hua/model/Criteria.java
@@ -10,7 +10,9 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.OneToMany;
-
+/**
+ * Class for the Criteria variables
+ */
 @Entity
 public class Criteria {
 	@Id
