+++ app-maven/src/main/java/com/hua/model/GenerateAhpResultsExcell.java
@@ -1,23 +1,0 @@
-package com.hua.model;
-
-import java.util.ArrayList;
-
-public class GenerateAhpResultsExcell {
-	private ArrayList<GenerateAhpResultsExcellWithSelection> generateAhpResultsExcellWithSelection;
-	private int id;
-
-	public ArrayList<GenerateAhpResultsExcellWithSelection> getGenerateAhpResultsExcellWithSelection() {
-		return generateAhpResultsExcellWithSelection;
-	}
-	public void setGenerateAhpResultsExcellWithSelection(
-			ArrayList<GenerateAhpResultsExcellWithSelection> generateAhpResultsExcellWithSelection) {
-		this.generateAhpResultsExcellWithSelection = generateAhpResultsExcellWithSelection;
-	}
-	public int getId() {
-		return id;
-	}
-	public void setId(int id) {
-		this.id = id;
-	}
-	
-}
