+++ app-maven/src/main/java/com/hua/model/AlternativesCriteriaCrAhp.java
@@ -7,7 +7,10 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.OneToOne;
-
+/**
+ * the Alternatives Criteria answers for the AHP algorithm
+ * it save the CR
+ */
 @Entity
 public class AlternativesCriteriaCrAhp {
 	@Id
