### 2022-10-11 0b79eec
Registration test for beans
