+++ app-maven/src/main/java/com/hua/util/GenerateAhpExcel.java
@@ -1,0 +1,283 @@
+package com.hua.util;
+
+import java.io.ByteArrayInputStream;
+import java.io.ByteArrayOutputStream;
+import java.io.IOException;
+import java.util.List;
+import java.util.Optional;
+
+import org.apache.poi.ss.usermodel.*;
+import org.apache.poi.ss.util.CellRangeAddress;
+import org.apache.poi.xssf.usermodel.*;
+import org.springframework.core.io.InputStreamResource;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+
+import com.hua.model.Alternatives;
+import com.hua.model.AlternativesCriteriaAnswerAhp;
+import com.hua.model.AlternativesFactorAnswerAhp;
+import com.hua.model.CriteriaAnswerAhp;
+import com.hua.model.FactorAnswerAhp;
+import com.hua.model.GenerateAhpResultsExcel;
+import com.hua.model.GenerateAhpResultsExcelWithSelection;
+import com.hua.model.ProblemUserAhp;
+import com.hua.repository.ProblemsUserAhpRepository;
+
+public class GenerateAhpExcel {
+	public ResponseEntity<InputStreamResource> makeExcel(int id, ProblemsUserAhpRepository problemsUserAhpRepository,
+														 GenerateAhpResultsExcel generateAhpResultsExcel){
+		Optional<List<ProblemUserAhp>> problemUserAhpListOpt = problemsUserAhpRepository.findAllByProblemUserProblemId(id);
+
+		if ( problemUserAhpListOpt.isPresent() ) {
+			List<ProblemUserAhp> problemUserAhpList = problemUserAhpListOpt.get();
+			Workbook workbook = new XSSFWorkbook();
+			workbook.createSheet("Persons");
+			for ( ProblemUserAhp problemUserAhp : problemUserAhpList ) {
+				
+				GenerateAhpResultsExcelWithSelection generateAhpExcel = generateAhpResultsExcel.getGenerateAhpResultsExcelWithSelection().stream()
+						  .filter(generateAhpResultsExcelWithSelectionTmp -> generateAhpResultsExcelWithSelectionTmp.getUser().getUsername().equals(problemUserAhp.getProblemUser().getUser().getUsername()))
+						  .findFirst()
+						  .orElse(null);
+				
+				if ( generateAhpExcel!=null && generateAhpExcel.getStatus()==true ) {
+					Sheet sheet = workbook.createSheet(problemUserAhp.getProblemUser().getUser().getUsername());
+					makeExceelPerUser(problemUserAhp, sheet);
+				}
+			}
+			makeExceelSum(workbook, problemUserAhpList, generateAhpResultsExcel);
+			
+			ByteArrayOutputStream out = new ByteArrayOutputStream();
+			try {
+				workbook.write(out);
+				
+				String filename = problemUserAhpList.get(0).getProblemUser().getProblem().getName().replace(" ", "_")+".xlsx";
+			    InputStreamResource file = new InputStreamResource(new ByteArrayInputStream(out.toByteArray()));
+
+			    return ResponseEntity.ok()
+			        .header(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=" + filename)
+			        .contentType(MediaType.parseMediaType("application/vnd.ms-excel"))
+			        .body(file);
+			} catch (IOException e) {
+				// TODO Auto-generated catch block
+				e.printStackTrace();
+			}
+		}
+		return null;
+		
+	}
[…]
