### 2022-10-11 e166bc2
Docker .env port improvements
Docker .env port improvements

Docker .env port improvements
