### 2022-10-17 38fbcc0
Add environmental simple user default credentials
