### 2022-10-17 gdede-hua/Decision_Survey_Platform#6 @gdede-hua
Dear @gcattan the ports are now included in the readme file.

The description on how to use the platform both as an administrator and as a user (how to add data) is also described in the readme file. You may also find the credentials for the default accounts of administrator and user.
