### 2022-10-17 gdede-hua/Decision_Survey_Platform#2 @gdede-hua
Dear @gcattan please be kindly informed that we have performed some JUnit tests included in the test folder.
