### 2022-10-17 gdede-hua/Decision_Survey_Platform#2 @gcattan
Ok, thank you for the update. I believe the work is now compliant with the JOSS criterion.
As a side comment (I leave it to you), I think it could be a good idea to run these tests into a pipeline.
