### 2022-10-19 gdede-hua/Decision_Survey_Platform#3 @gdede-hua
Dear @rreinecke thank you for your feedback. The paper.md has been updated as per your suggestion and now includes a  real case as an example.
