+++ Docs/Contributing.md
@@ -1,5 +1,5 @@
 
-# Welcome to GitHub docs contributing guide <!-- omit in toc -->
+# Welcome to Decision Survey Platform contributing guide <!-- omit in toc -->
 
 Thank you for investing your time in contributing to our project! Any contribution you make will be reflected on [docs.github.com](https://docs.github.com/en) :sparkles:. 
 
