### 2022-10-19 gdede-hua/Decision_Survey_Platform#5 @gdede-hua
Dear @rreinecke please be kindly informed that a contributing.md file is now included in our repository.
