### 2022-10-19 gdede-hua/Decision_Survey_Platform#4 @gdede-hua
Dear @rreinecke thank you for your comments.
We have included the code information in the readme.md (https://github.com/gdede-hua/Decision_Survey_Platform#code-informations).
Moreover a contributing.md file has been created in order to provide the guidance (https://github.com/gdede-hua/Decision_Survey_Platform/blob/master/Docs/Contributing.md).
Finally, detailed comments have been added in the actual code.
