### 2022-10-21 gdede-hua/Decision_Survey_Platform#4 @rreinecke
Great that you have included that information!

Concerning the code I could not find any substantial improvements of the code documentation on the main branch. Am I on the wrong branch?

For example https://github.com/gdede-hua/Decision_Survey_Platform/blob/master/app-maven/src/main/java/com/hua/util/RunAHP.java to my understanding implements the AGP algorithm and nothing is documented.
