### 2022-10-21 #4403 @rreinecke
> @rreinecke please check whether the changes in the repository adress your concerns and add the missing checkmarks in your list if they do.

They have been partially addressed. The code documentation is still very much below the minimum amount I would expect to understand the functionality of the code.
