### 2022-10-21 gdede-hua/Decision_Survey_Platform#3 @rreinecke
Great! I am not sure the paper.md is the right place (because it is quite extensive - which is good! Really helped me to understand). I am ok leaving it there if the JOSS editor agrees. Otherwise I would suggest moving it to you user documentation or readme file.
