### 2022-10-21 gdede-hua/Decision_Survey_Platform#8 @rreinecke
Ignorant usage leads to errors
The following cases lead to errors. These should have been tested through automated test and should not lead to a crash of the application.

- [x] Starting a research (why isn't that called survey as well btw.?) without an existing user group.
- [x] Creating a user group without selecting any users.
- [x] Not part of the review but still a suggestion: it would be good if you can end a survey process before the set date has passed. Currently it is not possible to make any modifications as soon as it is running.
