### 2022-10-21 gdede-hua/Decision_Survey_Platform#8 @rreinecke
Also it is possible to set end dates in the past and add surveys with empty names.
This doesn't seem well tested to me.
Still this doesn't lead to being able to verify the results. How can I test your algorithm without having to wait now?
