### 2022-10-21 700c452
simplify and functionality documentation
simplify and functionality documentation
