+++ app-maven/src/main/java/com/hua/util/GenerateAhpExcel.java
@@ -25,6 +25,10 @@ import com.hua.model.ProblemUserAhp;
 import com.hua.repository.ProblemsUserAhpRepository;
 
 public class GenerateAhpExcel {
+
+	/**
+	 * Make the Excel file and start to insert the results of the AHP algorithm
+	 */
 	public ResponseEntity<InputStreamResource> makeExcel(int id, ProblemsUserAhpRepository problemsUserAhpRepository,
 														 GenerateAhpResultsExcel generateAhpResultsExcel){
 		Optional<List<ProblemUserAhp>> problemUserAhpListOpt = problemsUserAhpRepository.findAllByProblemUserProblemId(id);
@@ -40,12 +44,12 @@ public class GenerateAhpExcel {
 						  .findFirst()
 						  .orElse(null);
 				
-				if ( generateAhpExcel!=null && generateAhpExcel.getStatus()==true ) {
+				if ( generateAhpExcel!=null && generateAhpExcel.getStatus()) {
 					Sheet sheet = workbook.createSheet(problemUserAhp.getProblemUser().getUser().getUsername());
-					makeExceelPerUser(problemUserAhp, sheet);
+					makeExcelPerUser(problemUserAhp, sheet);
 				}
 			}
-			makeExceelSum(workbook, problemUserAhpList, generateAhpResultsExcel);
+			makeExcelSum(workbook, problemUserAhpList, generateAhpResultsExcel);
 			
 			ByteArrayOutputStream out = new ByteArrayOutputStream();
 			try {
@@ -66,17 +70,21 @@ public class GenerateAhpExcel {
 		return null;
 		
 	}
-	public void makeExceelSum(Workbook workbook, List<ProblemUserAhp> problemUserAhpList, GenerateAhpResultsExcel generateAhpResultsExcell) {
+
+	/**
+	 * Generate the tab with the average of all user data
+	 */
+	public void makeExcelSum(Workbook workbook, List<ProblemUserAhp> problemUserAhpList, GenerateAhpResultsExcel generateAhpResultsExcel) {
 		Sheet sheet = workbook.getSheet("Persons");
 		int countOfSheets=problemUserAhpList.size();
 		
 		Sheet sheetUser = null;
 		for ( ProblemUserAhp problemUserAhp : problemUserAhpList ) {
-			GenerateAhpResultsExcelWithSelection generateAhpExcel = generateAhpResultsExcell.getGenerateAhpResultsExcelWithSelection().stream()
+			GenerateAhpResultsExcelWithSelection generateAhpExcel = generateAhpResultsExcel.getGenerateAhpResultsExcelWithSelection().stream()
 					  .filter(generateAhpResultsExcelWithSelectionTmp -> generateAhpResultsExcelWithSelectionTmp.getUser().getUsername().equals(problemUserAhp.getProblemUser().getUser().getUsername()))
 					  .findFirst()
 					  .orElse(null);
-			if ( generateAhpExcel!=null && generateAhpExcel.getStatus()==true ) {
+			if ( generateAhpExcel!=null && generateAhpExcel.getStatus()) {
 				sheetUser = workbook.getSheet(problemUserAhp.getProblemUser().getUser().getUsername());
 				break;
 			}
@@ -86,7 +94,7 @@ public class GenerateAhpExcel {
 		int rowCounter=100, cellCounter=20;
 		for(int i=0; i<rowCounter; i++) {
 			for(int j=0; j<cellCounter; j++) {
-				Row headerUser = sheetUser.getRow(i);
+				Row headerUser = sheetUser != null ? sheetUser.getRow(i) : null;
 				if ( headerUser!=null ) {
 					Cell headerCellUser = headerUser.getCell(j);
[…]
