+++ app-maven/src/main/java/com/hua/util/RunAHP.java
@@ -36,8 +36,8 @@ import Jama.Matrix;
 @Service
 public class RunAHP {
 
-	private ProblemsUserRepository problemsUserRepository;
-	private ProblemsUserAhpRepository problemsUserAHPRepository;
+	private final ProblemsUserRepository problemsUserRepository;
+	private final ProblemsUserAhpRepository problemsUserAHPRepository;
 
 	public RunAHP(ProblemsUserRepository problemsUserRepository, ProblemsUserAhpRepository problemsUserAHPRepository) {
 		this.problemsUserRepository = problemsUserRepository;
@@ -57,188 +57,250 @@ public class RunAHP {
 				ProblemUserAhp problemUserAhp = new ProblemUserAhp();
 				problemUserAhp.setProblemUser(new ProblemUser(problemUser.getId(), problemUser.getUser()));
 				List<Criteria> criteriaList = problemUser.getProblem().getCriteria();
-				if (criteriaList.size() > 0) {
 
-					System.out.println(
-							"--------------------------------- Run Criteria  -----------------------------------");
-					double[][] matrixCriteria = new double[criteriaList.size()][criteriaList.size()];
-					for (int i = 0; i < criteriaList.size(); i++) {
-						for (int j = 0; j < criteriaList.size(); j++) {
-							if (i != j) {
-								final int iFinal = i;
-								final int jFinal = j;
-								CriteriaAnswer criteriaAnswer = problemUser.getCriteriaAnswer().stream()
-										.filter(criteriaAnswerTmp -> criteriaAnswerTmp.getCriteriaTop()
-												.getId() == criteriaList.get(iFinal).getId()
-												&& criteriaAnswerTmp.getCriteriaLeft().getId() == criteriaList
-														.get(jFinal).getId())
-										.findAny().orElse(null);
-								if (criteriaAnswer != null)
-									matrixCriteria[i][j] = criteriaAnswer.getWeight();
-								else
-									matrixCriteria[i][j] = 1;
-							} else
-								matrixCriteria[i][j] = 1;
-						}
-					}
-					AHP ahpDataCriteria = runAHPeig(matrixCriteria);
-					List<CriteriaAnswerAhp> criteriaAnswerAHPList = new ArrayList<>();
-					for (int i = 0; i < criteriaList.size(); i++) {
-						criteriaAnswerAHPList.add(new CriteriaAnswerAhp(new Criteria(criteriaList.get(i).getId()),
-								(float) ahpDataCriteria.getWeights()[i]));
-					}
-					problemUserAhp.setCriteriaCrAhp(new CriteriaCrAhp((float) ahpDataCriteria.getCr()));
-					problemUserAhp.setCriteriaAnswerAhp(criteriaAnswerAHPList);
-				}
-				System.out.println(
-						"--------------------------------- Complete Criteria  -----------------------------------");
+				criteriaAhp(problemUser, problemUserAhp, criteriaList);
 
-				List<FactorAnswerAhp> factorAnswerAHPList = new ArrayList<>();
-				List<FactorCrAhp> factorCrAhpList = new ArrayList<>();
-				for (Criteria criteria : criteriaList) {
-					if (criteria.getFactor().size() > 1) {
-						System.out.println("--------------------------------- Run Facto for Criteria: "
-								+ criteria.getName() + "  -----------------------------------");
-						List<Factor> factorList = criteria.getFactor();
-						double[][] matrixFactor = new double[factorList.size()][factorList.size()];
-						for (int i = 0; i < factorList.size(); i++) {
-							for (int j = 0; j < factorList.size(); j++) {
-								if (i != j) {
-									final int iFinal = i;
-									final int jFinal = j;
[…]

[…]
