### 2022-10-24 gdede-hua/Decision_Survey_Platform#4 @fabian-s
I tend to agree with @rreinecke here, not a lot of very useful documentation is present/has been added in the current master since April.
JOSS requires at least a complete API method documentation, which doesn't seem to exist here yet.

@gdede-hua: please also note that these reviews become a lot simpler for us reviewers/editors if your commit messages explicitly reference the respective issues they are meant to fix, or if your messages in these issue threads explicitly reference the commits that were intended to fix them.
