### 2022-10-24 gdede-hua/Decision_Survey_Platform#8 @fabian-s
@gede-hua to me these seem like valid criticisms re. the functionality and stability of the software you'd like JOSS to publish and the quality of automated testing in your code. Please engage with these issues.

please also note that these reviews become a lot simpler for us reviewers/editors if your commit messages explicitly reference the respective issues they are meant to fix, or if your messages in these issue threads explicitly reference the commits that were intended to fix them.
