### 2022-10-24 gdede-hua/Decision_Survey_Platform#3 @fabian-s
@gdede-hua JOSS aims for papers <1000 words, you're now at  ~1900.
Please move the case study out of the paper and reference it there.

see also #9
