### 2022-10-24 gdede-hua/Decision_Survey_Platform#9 @fabian-s
paper is too long for JOSS
JOSS aims for papers <1000 words, you're now at  ~1900.
Please move the case study out of the paper into a file hosted in your repo, and reference it in your paper. (no need to remove anything else)

see also https://github.com/openjournals/joss-reviews/issues/4403
