### 2022-11-01 gdede-hua/Decision_Survey_Platform#8 @gdede-hua
Dears,
Thank you for your comments.
Please be kindly informed that additional html validations are now included in order address the raised issues by @rreinecke. You may find the commit here: https://github.com/gdede-hua/Decision_Survey_Platform/commit/cc829099be6181fb744b327bb042dcc64abdb164.

Moreover, a complete survey button has been added in order to get the results and not waiting the end of date.
Finally, "Research" has been replaced by "survey".
