### 2022-11-01 gdede-hua/Decision_Survey_Platform#4 @gdede-hua
Dears, may I kindly ask you to provide me more information about what you expect in the API method documentation?
