### 2022-11-02 gdede-hua/Decision_Survey_Platform#9 @gdede-hua
Dear @fabian-s please be informed that the case study has been removed from the paper and it is now included in readme.
