### 2022-11-02 gdede-hua/Decision_Survey_Platform#3 @gdede-hua
Dears, please be kindly informed that the case study has been moved to the readme file.
