### 2022-11-14 gdede-hua/Decision_Survey_Platform#5 @fabian-s
The convention is for CONTRIBUTING.md to be in the top level folder of the repo, not in /Docs. Please move it there.
