### 2022-11-14 gdede-hua/Decision_Survey_Platform#4 @fabian-s
@gdede-hua: well, precise and correct descriptions of the classes you define and their methods need to be accessible in some way, at least for the user-facing ones. AFAICT, your code remains mostly undocumented.

@rreinecke could you provide some more specific technical advice on how to do this the right way?
I don't program in Java myself.
