### 2022-11-14 gdede-hua/Decision_Survey_Platform#3 @fabian-s
@rreinecke please close this issue if you are satisfied with the example
