### 2022-11-14 gdede-hua/Decision_Survey_Platform#3 @rreinecke
the examples contains a typo in line three with one point too much.

Apart from that can a potential user download that scenario and look at the results? Then you could shorten the text a bit and users could try it out themself. I think that would be helpful. Also at least one screenshot of how the results look like would be good. Currently it is outline an example but does not really demonstrate on how your tool then processes the described example and how that would be helpful. Or is the tool really just about collecting the decisions? If so then I think a potential result figure would still be helpful I think.
