### 2022-11-14 gdede-hua/Decision_Survey_Platform#4 @rreinecke
I think this blog provides a nice overview: https://www.jrebel.com/blog/tips-and-tricks-for-better-java-documentation
