### 2022-11-15 gdede-hua/Decision_Survey_Platform#8 @fabian-s
@rreinecke are you satisfied with these changes? if so, please close this issue.
if not, please make a note what is still missing here.
