### 2022-11-15 gdede-hua/Decision_Survey_Platform#5 @fabian-s
Your Contributing.md file seems to be a verbatim copy of https://github.com/github/docs/blob/3203ecbc31766631f3bfa093daa0f4d1bcbb0564/CONTRIBUTING.md.

This is useless "going through the motions" on your part - many of the remarks in there are not applicable or sensible for this repository and and do not inform people about how *you* would expect them to collaborate on *this*.
Please adapt the text to the specifics of this project.
