### 2022-11-17 gdede-hua/Decision_Survey_Platform#10 @fabian-s
Automated tests
https://github.com/openjournals/joss-reviews/issues/4403

@rreinecke
this commit:
https://github.com/gdede-hua/Decision_Survey_Platform/commit/5fadffb8263bfc23c4da80735037a6bc5f166e25
added/extended unit tests AFAICT -- do these satisfy your concerns re. the lack of automated testing?
