### 2022-11-21 gdede-hua/Decision_Survey_Platform#10 @rreinecke
I think it could still be improved but it is certainly  start. Thank you for implementing this.
If you like close this issue.
