### 2022-11-22 gdede-hua/Decision_Survey_Platform#5 @fabian-s
@gdede-hua
please add a reasonably project-specific CONTRIBUTING.md file to the top level folder of your repo and reply with a reference to the relevant commit within the next 7 days so we can finally wrap up this process.
