### 2023-01-17 gdede-hua/Decision_Survey_Platform#4 @gdede-hua
@fabian-s please be kindly informed that I will address all the comments by end of January. Thank you.
