### 2023-03-10 gdede-hua/Decision_Survey_Platform#5 @skatz91
Dear @fabian-s
[commit](https://github.com/gdede-hua/Decision_Survey_Platform/commit/f8f5a3b93a5a6fdabf501f80912c96b9213df294)

The CONTRIBUTING.md is now more project specific. Let me know if we have more problems with the file.

Best regards
John Nikolaou
