### 2023-03-10 gdede-hua/Decision_Survey_Platform#4 @skatz91
Dear @fabian-s & @rreinecke
[commit](https://github.com/gdede-hua/Decision_Survey_Platform/commit/c8b0d6e3afb42301068fdafa7ddaff75542ab3c4)

I have added a lot of information for our function. Could you let me know if you want more changes?

Best regards
John Nikolaou
