### 2023-03-20 gdede-hua/Decision_Survey_Platform#4 @fabian-s
These are not sufficiently useful or informative comments for  documenting an API in my opinion-- see the very first comment by @rreinecke in this issue on the standard of documentation expected for JOSS-published software.
