### 2023-03-20 gdede-hua/Decision_Survey_Platform#5 @fabian-s
Still seems like useless "going through the motions" on your part - many of the remarks in there are not applicable or sensible for this repository and and do not inform people about how you would expect them to collaborate on this.
Please adapt the text to the specifics of this project, beyond exchanging a couple of invalid URLs that you mindlessly copied over.....
