### 2023-07-11 #4403 @fabian-s
@openjournals/joss-eics sorry for not taking better care of this submission.

The authors have not responded to requests to fix outstanding issues for months now. My recommendation is to reject this submission since it does not provide community guidelines nor achieve any reasonable standard for API and code documentation.
